requirements.txt
djangorestframework==3.14.0
drf_spectacular==0.26.2
